+++ EoRadio/app/src/main/java/dk/dr/radio/afspilning/wrapper/MediaPlayerWrapper.java
@@ -20,11 +20,11 @@ public interface MediaPlayerWrapper {
 
   void release();
 
-  void seekTo(int offsetMs);
+  void seekTo(long offsetMs);
 
-  int getDuration();
+  long getDuration();
 
-  int getCurrentPosition();
+  long getCurrentPosition();
 
   void start();
